Reader: Gcal EventFile setting & cleanup.
- Most of these changes are purely cosmetic; I want to make the script a
little more human-readable before releasing it as the next Universal
Feed Reader update.

- I consolidated the Google Calendar stuff so that it can now be spliced
into the main script with only two lines.

- Event files are now defined as a Script measure setting. I removed the
function that checked #CalendarEventFile# and added the Gcal file to it;
it was clever, but I don't want to change user variables without their
permission. It's enough that the file is written and available.