fixed #6
With the "isdst=false" assumption removed, the date from Google Calendar
will be converted as DST or not-DST according to the current system
setting. The only downside is that events on the other side of the line
in spring/fall will be wrong. For a more comprehensive fix, we may need
to calculate whether a date falls within DST on our own.